Create methods of Math---pow() sqrt() and cbrt()
Coding in function cutCube. function accept 2 parameter: volume and n. volume is the volume of a cube. If we cut the cube into n block. please determine whether the length of the cube is an integer. return true or false.
For exmaple:
volume=27, it can be divided into 27 blocks, each small cube side length is 1